// src/app/api/model/createDespesasDto.ts
/**
 * DPaulaCont
 * Descrição - DPaulaCont API 
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CreateDespesasDto { 
    fonte: number;
    data: Date;
    destinado: string;
    valor: number;
    pagoPara: number;
    centroCusto: number;
}